kingpin/pyproject.toml
[tool.poetry]
name = "kingpin"
version = "0.1.0"
description = "The main controller API of the Goblin Wrangler"
authors = ["Stephen Badger <stephen.badger@gmail.com>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
gunicorn = "^20.0.4"
flask-login = "^0.5.0"
loginpass = "^0.5"
authlib = "^0.14.3"
psycopg2-binary = "^2.8.5"
Flask-SQLAlchemy = "^2.4.4"
flask-migrate = "^2.5.3"
flask-script = "^2.0.6"

[tool.poetry.dev-dependencies]
pytest = "^6.0.1"
pytest-cov = "^2.10.0"
pylint = "^2.5.3"
pycodestyle = "^2.6.0"

[tool.poetry.scripts]
manage = "kingpin.db.manage:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
